test_fourfront_utils tests fully migrated in snovault
